refactor(패키지명 변경): beside.xxx.xxxserver -> beside.sunday8turtle.pickabook 으로 패키지명 변경
