+++ src/test/java/beside/sunday8turtle/pickabookserver/XxxServerApplicationTests.java
@@ -1,4 +1,4 @@
-package beside.xxx.xxxserver;
+package beside.sunday8turtle.pickabookserver;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
